chore: Add renamed files to build
