+++ src/CMakeLists.txt
@@ -1,2 +1,2 @@
-add_library(source inverted_pendulum.cpp tools.cpp)
+add_library(source inverted_pendulum.cpp tools.cpp pid.cpp)
 target_include_directories(source PUBLIC $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}>)
